src/requirements.txt
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
SQLAlchemy==2.0.29
gunicorn==21.2.0
azure-identity==1.16.0
